+++ packages/components/src/DateTimePickers/InputTimePicker/InputTimePicker.component.js
@@ -30,6 +30,7 @@ export default function InputTimePicker(props) {
 	const containerRef = useRef(null);
 
 	const handlers = useInputPickerHandlers({
+		disabled: props.disabled,
 		handleBlur: props.onBlur,
 		handleChange: props.onChange,
 		handleKeyDown: () => focusOnTime(containerRef.current),
@@ -37,12 +38,7 @@ export default function InputTimePicker(props) {
 
 	const inputProps = omit(props, PROPS_TO_OMIT_FOR_INPUT);
 	const timePicker = [
-		<Time.Input
-			{...inputProps}
-			id={`${props.id}-input`}
-			key="input"
-			inputRef={inputRef}
-		/>,
+		<Time.Input {...inputProps} id={`${props.id}-input`} key="input" inputRef={inputRef} />,
 		handlers.showPicker && (
 			<Popper
 				key="popper"
@@ -99,6 +95,7 @@ InputTimePicker.propTypes = {
 	onBlur: PropTypes.func,
 	timezone: PropTypes.string,
 	value: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+	disabled: PropTypes.bool,
 };
 
 InputTimePicker.defaultProps = {