+++ packages/components/src/FilterBar/FilterBar.test.js
@@ -43,9 +43,38 @@ describe('FilterBar', () => {
 
 	it('should accept data-test attribute', () => {
 		// given
-		const filterInstance = mount(<FilterBarComponent {...defaultProps} data-test="my.test" />);
+		const filterInstance = mount(
+			<FilterBarComponent {...defaultProps} data-test="my.test" value="filter text" />,
+		);
 		// then
 		expect(filterInstance.find('input').prop('data-test')).toEqual('my.test');
+	});
+
+	it('should accept data-feature attribute', () => {
+		// given
+		const filterInstance = mount(
+			<FilterBarComponent {...defaultProps} data-feature="my.feature" />,
+		);
+		// then
+		expect(filterInstance.find('input').prop('data-feature')).toEqual('my.feature');
+	});
+
+	it('should propagate and compose data-test attribute to the clear selection button', () => {
+		// given
+		const filterInstance = mount(
+			<FilterBarComponent {...defaultProps} data-test="my.test" value="filter text" focus />,
+		);
+		// then
+		expect(filterInstance.find('button').prop('data-test')).toEqual('my.test-reset');
+	});
+
+	it('should propagate and compose data-feature attribute to the clear selection button', () => {
+		// given
+		const filterInstance = mount(
+			<FilterBarComponent {...defaultProps} data-test="my.test" value="filter text" focus />,
+		);
+		// then
+		expect(filterInstance.find('button').prop('data-test')).toEqual('my.test-reset');
 	});
 
 	it('should be able to switch autofocus to false', () => {