+++ packages/dataviz/src/types/index.ts
@@ -4,11 +4,6 @@ export type Range = {
 	excludeMax?: boolean;
 };
 
-export enum DataType {
-	DATE = 'DATE',
-	NUMBER = 'NUMBER',
-}
-
 export enum ChartStyle {
 	VALUE = 'value',
 	PATTERN = 'pattern',